Modularize chat view and models; move logic to `chat_view.rs` and `models.rs` for improved code organization and maintainability.
